requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
auth0-python==4.9.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
crewai==0.114.0
crewai-tools==0.40.1
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
embedchain==0.1.128
et_xmlfile==2.0.0
executing==0.8.3
fastapi==0.115.9
fastavro==1.10.0
fastjsonschema==2.20.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.2
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
instructor==1.7.9
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json_repair==0.41.1
json5==0.12.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.0
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==32.0.1
lancedb==0.21.2
langchain==0.3.23
langchain-cohere==0.3.5
langchain-community==0.3.21
langchain-core==0.3.54
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.32
litellm==1.60.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mem0ai==0.1.92
mistune==3.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pip==25.0
platformdirs==4.3.7
portalocker==2.10.1
posthog==3.25.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.9.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.399
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.3
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
schema==0.7.7
Send2Trash==1.8.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.40
stack-data==0.2.0
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
terminado==0.17.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-requests==2.32.0.20250328
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uv==0.6.14
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
